Separate RocksDB checkpoint implementation from server core
