+++ fdbserver/checkpoint/include/fdbserver/checkpoint/Checkpoint.h
@@ -1,0 +1,45 @@
+/*
+ * Checkpoint.h
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2026 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+#include "fdbserver/core/ServerCheckpoint.h"
+
+ICheckpointReader* newCheckpointReader(const CheckpointMetaData& checkpoint,
+                                       const CheckpointAsKeyValues checkpointAsKeyValues,
+                                       UID logID);
+
+// Delete a checkpoint.
+Future<Void> deleteCheckpoint(CheckpointMetaData checkpoint);
+
+// Fetches checkpoint to a local `dir`, `initialState` provides the checkpoint formats, location, restart point, etc.
+// If cFun is provided, the progress can be checkpointed.
+// Returns a CheckpointMetaData, which could contain KVS-specific results, e.g., the list of fetched checkpoint files.
+Future<CheckpointMetaData> fetchCheckpoint(Database cx,
+                                           CheckpointMetaData initialState,
+                                           std::string dir,
+                                           std::function<Future<Void>(const CheckpointMetaData&)> cFun = nullptr);
+
+// Same as above, except that the checkpoint is fetched as key-value pairs.
+Future<CheckpointMetaData> fetchCheckpointRanges(Database cx,
+                                                 CheckpointMetaData initialState,
+                                                 std::string dir,
+                                                 std::vector<KeyRange> ranges,
+                                                 std::function<Future<Void>(const CheckpointMetaData&)> cFun = nullptr);
