+++ fdbserver/workloads/CMakeLists.txt
@@ -14,6 +14,7 @@ target_include_directories(fdbserver_workloads PRIVATE
 target_link_libraries(fdbserver_workloads PRIVATE
   fdbserver_consistencyscan
   fdbserver_core
+  fdbserver_checkpoint
   fdbserver_kvstore
   fdbserver_worker
   fdbserver_tester
@@ -21,3 +22,7 @@ target_link_libraries(fdbserver_workloads PRIVATE
   fdbserver_resolver
   fdbserver_storageserver
   fdbserver_mocks3)
+
+if(WITH_ROCKSDB)
+  target_compile_definitions(fdbserver_workloads PRIVATE WITH_ROCKSDB)
+endif()
