+++ fdbserver/workloads/StorageServerCheckpointRestoreTest.cpp
@@ -23,7 +23,7 @@
 #include "fdbrpc/simulator.h"
 #include "fdbserver/kvstore/IKeyValueStore.h"
 #include "fdbserver/core/Knobs.h"
-#include "fdbserver/core/ServerCheckpoint.h"
+#include "fdbserver/checkpoint/Checkpoint.h"
 #include "fdbserver/core/MoveKeys.h"
 #include "fdbserver/core/QuietDatabase.h"
 #include "fdbserver/tester/workloads.h"
